pyproject.toml
[project]
name = "schwab-mcp"
version = "0.1.0"
description = "Schwab Model Context Protocol (MCP) for the Schwab API"
readme = "README.md"
authors = [
    { name = "Jason Kölker", email = "jason@koelker.net" }
]
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.9.0",
    "click>=8.1.8",
    "mcp>=1.4.1",
    "platformdirs>=4.3.7",
    "pyyaml>=6.0.2",
]

[project.scripts]
schwab-mcp = "schwab_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mcp[cli]>=1.4.1",
    "pyright>=1.1.397",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]
